feat: rename course edit page to settings page
