+++ app/(main)/courses/[course_id]/settings/admin/page.tsx
@@ -1,0 +1,3 @@
+export default function CourseSettingsAdminPage() {
+  return <>Admin</>;
+}
